Fixed a warning

+++ DACircularProgress/DACircularProgressView.m
@@ -154,7 +154,7 @@
     if (animated)
     {
         CABasicAnimation *animation = [CABasicAnimation animationWithKeyPath:@"progress"];
-        animation.duration = fabsf(self.progress - pinnedProgress) * self.durationFactor; // Same duration as UIProgressView animation
+        animation.duration = fabs(self.progress - pinnedProgress) * self.durationFactor; // Same duration as UIProgressView animation
         animation.timingFunction = [CAMediaTimingFunction functionWithName:kCAMediaTimingFunctionEaseInEaseOut];
         animation.fromValue = [NSNumber numberWithFloat:self.progress];
         animation.toValue = [NSNumber numberWithFloat:pinnedProgress];
